rename ssl to self_supervised_learning, add readme
