+++ examples/self_supervised_learning/data_modules/__init__.py
