+++ examples/self_supervised_learning/lightning.py
