+++ examples/self_supervised_learning/losses/__init__.py
