+++ examples/self_supervised_learning/lr_schedulers/__init__.py
@@ -1,0 +1,5 @@
+from ._linear_decay import LinearDecayLRScheduler
+
+__all__ = [
+    "LinearDecayLRScheduler",
+]
