+++ examples/self_supervised_learning/lr_schedulers/_linear_decay.py
