+++ examples/self_supervised_learning/train_hubert.py
@@ -1,18 +1,23 @@
 import logging
 import pathlib
-from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser, RawDescriptionHelpFormatter
+from argparse import (
+    ArgumentDefaultsHelpFormatter,
+    ArgumentParser,
+    RawDescriptionHelpFormatter,
+)
 from functools import partial
 from typing import Dict, Tuple
 
 import torch
 import torchaudio.models
-from data_modules import HuBERTDataModule
-from lightning import SSLPretrainModule
-from losses import hubert_loss
-from lr_schedulers import LinearDecayLRScheduler
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.utilities.seed import seed_everything
+
+from .data_modules import HuBERTDataModule
+from .lightning import SSLPretrainModule
+from .losses import hubert_loss
+from .lr_schedulers import LinearDecayLRScheduler
 
 
 class _Formatter(ArgumentDefaultsHelpFormatter, RawDescriptionHelpFormatter):
