--- examples/ssl/lr_schedulers/__init__.py
@@ -1,5 +1,0 @@
-from _linear_decay import LinearDecayLRScheduler
-
-__all__ = [
-    "LinearDecayLRScheduler",
-]
